requirements.txt
Pillow>=9.0.0
numpy
opencv-python
pytesseract
pyautogui
playwright
selenium
